fix some NPE

+++ trunk/riverock-portlet/src/java/org/riverock/portlet/member/currency/CurrencyDateChangeClassQuery.java
@@ -122,7 +122,7 @@ public class CurrencyDateChangeClassQuery extends BaseClassQuery
 
         if (cat.isDebugEnabled())
 
-            cat.debug("idCurrency - "+idCurrency.longValue());
+            cat.debug("idCurrency - "+idCurrency);
 
     }
