+++ trunk/riverock-portlet/src/java/org/riverock/portlet/member/currency/CurrencyNameStdCurrencyClassQuery.java
@@ -120,7 +120,7 @@ public class CurrencyNameStdCurrencyClassQuery extends BaseClassQuery
 
         if (cat.isDebugEnabled())
 
-            cat.debug("idCurrency - "+idCurrency.longValue());
+            cat.debug("idCurrency - "+idCurrency);
 
     }
